Updated unit tests for input validation in new folder
