+++ tests/units/input_models/routing/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2023-2024 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+"""Test for anta.input_models.routing submodule."""
